requirements.txt
numpy==1.23.3
torch==1.12.1
rich==12.6.0
pillow==9.2.0
scipy==1.9.2
pandas==1.5.0
albumentations==1.3.0
torchvision==0.13.1
pytorch_lightning==1.7.7
stardist==0.8.3